pyproject.toml
[project]
name = "citation-arbitrage"
version = "0.1.0"
description = "Arbitrage academic citation impact to identify overlooked startup opportunities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "networkx>=3.0",  # For PageRank computation
    "beautifulsoup4>=4.12.0",  # For web scraping
    "selenium>=4.0.0",  # For LinkedIn scraping
]

[project.scripts]
citation-arbitrage = "citation_arbitrage.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
